requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-environ==0.11.2
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
gunicorn==23.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
